Dataset_creation/requirements.txt
openai>=1.68.0
elevenlabs>=2.14.0
google-genai>=0.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
pydub>=0.25.0
pathlib>=1.0.0
asyncio-throttle>=1.0.0
